Refactored deploy-commands to a method within BotClient, now pushCommands
