+++ src/commands/utility/commands.js
@@ -1,10 +1,13 @@
 const { SlashCommandBuilder } = require('discord.js');
 
+
+
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('commands')
 		.setDescription('Lists all avalible commands!'),
 	async execute(interaction) {
-		await interaction.reply('test');
+		await interaction.reply(`/commands,
+			/ping`);
 	},
 };
